mmap test

+++ src/handlers.zig
@@ -23,10 +23,11 @@ inline fn epilogue(data_ptr: u64, accum: u64) void {
 pub fn readHandler() callconv(.C) void {
     const data_ptr = azm.getDataPtr();
     const accum = azm.getAccum();
-    const write_to: [*]u8 = @ptrFromInt(data_ptr);
+    const write_to: [*]u64 = @ptrFromInt(data_ptr);
     writeBufferFlush();
     const stdin = std.io.getStdIn().reader();
     _ = stdin.read(write_to[0..1]) catch @panic("READ FAILED");
+
     epilogue(data_ptr, accum);
 }
